refactor(build_transformer): move artefact download ops
